// src/lib/types.ts
export interface PropertyFilter {
  propertyName: string;
  dataType: number;
  operator: Operator;
  value: string;
}

export type EventPropertyFilter = PropertyFilter;
export type SessionPropertyFilter = PropertyFilter;
// ...
export interface QueryFilters
  extends DateParams,
    FilterParams,
    SortParams,
    PageParams,
    SegmentParams {
  minDuration?: number;
  cohortFilters?: QueryFilters;
  eventPropertyFilters?: EventPropertyFilter[];
  sessionPropertyFilters?: SessionPropertyFilter[];
}

export interface DateParams {
  startDate?: Date;
  endDate?: Date;
  unit?: string;
  timezone?: string;
  compareDate?: Date;
}

export interface FilterParams {
  path?: string;
  referrer?: string;
  title?: string;
  query?: string;
  host?: string;
  os?: string;
  browser?: string;
  device?: string;
  country?: string;
  region?: string;
  city?: string;
  language?: string;
  event?: string;
  search?: string;
  tag?: string;
  eventType?: number;
  segment?: string;
  cohort?: string;
  compare?: string;
  excludeBounce?: boolean;
  match?: 'all' | 'any';
}

export interface SortParams {
  orderBy?: string;
  sortDescending?: boolean;
}

export interface PageParams {
  page?: number;
  pageSize?: number;
  maxResults?: number;
}

export interface SegmentParams {
  segment?: string;
  cohort?: string;
}